+++ src/livestreamer/plugins/afreecatv.py
@@ -2,29 +2,36 @@ import re
 
 from livestreamer.plugin import Plugin
 from livestreamer.plugin.api import http, validate
-from livestreamer.stream import HLSStream
+from livestreamer.stream import RTMPStream, HLSStream
 
-USER_AGENT = (
-    "Mozilla/5.0 (iPhone; CPU iPhone OS 6_0 like Mac OS X) AppleWebKit/536.26 "
-    "(KHTML, like Gecko) Version/6.0 Mobile/10A5376e Safari/8536.25"
-)
-HEADERS = {"User-Agent": USER_AGENT}
-PLAYLIST_URL = "http://m.afreeca.com/live/stream/a/hls/broad_no/{broad_no}"
-BROAD_INFO_URL = "https://api.m.afreeca.com/broad/a/getbroadinfo"
+CHANNEL_INFO_URL = "http://live.afreeca.com:8057/afreeca/player_live_api.php"
+KEEP_ALIVE_URL = "{server}/stream_keepalive.html"
+STREAM_INFO_URLS = {
+    "rtmp": "http://sessionmanager01.afreeca.tv:6060/broad_stream_assign.html",
+    "hls": "http://resourcemanager.afreeca.tv:9090/broad_stream_assign.html"
+}
 
-_broadcast_re = re.compile(r".+\/(\d+)\.gif")
+CHANNEL_RESULT_ERROR = 0
+CHANNEL_RESULT_OK = 1
+
+
 _url_re = re.compile("http(s)?://(\w+\.)?afreeca.com/(?P<username>\w+)")
 
-_broadcast_schema = validate.Schema(
+_channel_schema = validate.Schema(
     {
-        "data": validate.any(
-            None,
-            {
-                "broad_no": int,
-            }
-        )
+        "CHANNEL": {
+            "RESULT": validate.transform(int),
+            validate.optional("BNO"): validate.text,
+        }
     },
-    validate.get("data")
+    validate.get("CHANNEL")
+)
+_stream_schema = validate.Schema(
+    {
+        validate.optional("view_url"): validate.url(
+            scheme=validate.any("rtmp", "http")
+        )
+    }
 )
 
 
@@ -33,26 +40,52 @@ class AfreecaTV(Plugin):
     def can_handle_url(self, url):
         return _url_re.match(url)
 
-    def _get_broadcast(self, username):
+    def _get_channel_info(self, username):
         data = {
-            "szBjId": username
+            "bid": username
         }
-        res = http.post(BROAD_INFO_URL, headers=HEADERS, data=data)
+        res = http.post(CHANNEL_INFO_URL, data=data)
 
-        return http.json(res, schema=_broadcast_schema)
+        return http.json(res, schema=_channel_schema)
+
+    def _get_stream_info(self, broadcast, type):
+        params = {
+            "return_type": "gs_cdn",
+            "broad_key": "{broadcast}-flash-hd-{type}".format(**locals())
+        }
+        res = http.get(STREAM_INFO_URLS[type], params=params)
+        return http.json(res, schema=_stream_schema)
+
+    def _get_hls_stream(self, broadcast):
+        info = self._get_stream_info(broadcast, "hls")
+        if "view_url" in info:
+            return HLSStream(self.session, info["view_url"])
+
+    def _get_rtmp_stream(self, broadcast):
+        info = self._get_stream_info(broadcast, "rtmp")
+        if "view_url" in info:
+            params = dict(rtmp=info["view_url"])
+            return RTMPStream(self.session, params=params, redirect=True)
 
     def _get_streams(self):
         match = _url_re.match(self.url)
         username = match.group("username")
 
-        broadcast = self._get_broadcast(username)
+        channel = self._get_channel_info(username)
+        if channel["RESULT"] != CHANNEL_RESULT_OK:
+            return
+
+        broadcast = channel.get("BNO")
         if not broadcast:
             return
 
-        playlist_url = PLAYLIST_URL.format(**broadcast)
-        stream = HLSStream(self.session, playlist_url, headers=HEADERS)
+        flash_stream = self._get_rtmp_stream(broadcast)
+        if flash_stream:
+            yield "live", flash_stream
 
-        return dict(live=stream)
+        mobile_stream = self._get_hls_stream(broadcast)
+        if mobile_stream:
+            yield "live", mobile_stream
 
 
 __plugin__ = AfreecaTV